fixed go generate command to use proper package name and made benchmark use new tool
